Adding admin methods + separate class for creating objects with input from user

+++ src/AeronefBuilder.java
@@ -30,7 +30,7 @@ public class AeronefBuilder implements Builder{
     }
 
     @Override
-    public void setPriceKM(int priceKM) {
+    public void setPriceKM(double priceKM) {
         this.priceKM = priceKM;
     }
